+++ CHANGELOG.md
@@ -5,8 +5,7 @@
 * Fixed an issue where **validate** would incorrectly identify files as renamed.
 * Added a validation that integration properties (such as feed, mappers, mirroring, etc) are not removed.
 * Fixed an issue where **validate** failed when comparing branch against commit hash.
-* Added a validation that there's no change is incidents fields, or removal of incidents types
-* Fixed an issue where the *create-id-set* flag in the *validate* command did not work while not using git
+* Added the *--no-pipenv* flag to the **split-yml** command.
 
 # 1.3.0
 * Do not collect optional dependencies on indicator types reputation commands.